refactor(docs): fold check_doc_links logic into test_doc_links.py
The link checker's only consumer is its own test file; a dedicated
module under reflex_docs.scripts was extra indirection for no benefit.
Inline the helpers (`_normalize`, `_walk_blocks`, `check`, etc.) at
the top of tests/test_doc_links.py and delete
docs/app/reflex_docs/scripts/ entirely.
